Fix D3D12_DEPTH_STENCIL_VALUE serialization.
The character value was being encoded instead of the numeric value since UINT8 is also the unsigned char type used by ostream operator<<.

+++ src/gpgmm/d3d12/JSONSerializerD3D12.cpp
@@ -90,7 +90,7 @@ namespace gpgmm { namespace d3d12 {
         std::stringstream ss;
         ss << "{ "
            << "\"Depth\": " << depthStencilValue.Depth << ", "
-           << "\"Stencil\": " << depthStencilValue.Stencil << " }";
+           << "\"Stencil\": " << static_cast<uint32_t>(depthStencilValue.Stencil) << " }";
         return ss.str();
     }
